requirements.txt
pygame>=2.5.0
Flask>=3.1.0
mistralai
python-dotenv